Fixing failing Travis
