+++ apps/volontulo/tests/views/offers/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+u"""
+.. module:: __init__
+"""
