+++ apps/volontulo/views/organizations.py
@@ -3,6 +3,7 @@
 u"""
 .. module:: organizations
 """
+
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.core.urlresolvers import reverse
